# apps/api/src/swagger.yaml
openapi: 3.0.3
info:
  title: AgriTrack Central API
  description: |
    AgriTrack Central Backend API for Smart CRM Machinery Monitoring.
    
    ## Features
    - **Machines**: Real-time machine monitoring and management
    - **Bookings**: Farmer booking system
    - **Analytics**: Dashboard statistics
    - **Notifications**: SMS, Push, and Email alerts
    - **Geofencing**: Location-based boundaries and breach detection
    - **Fuel Tracking**: Real-time fuel monitoring and theft detection
    - **Maintenance**: Scheduled maintenance management
    
    ## WebSocket Events
    Connect to `ws://localhost:3001` for real-time updates:
    - `device_update`: Real-time machine state
    - `booking_update`: Booking status changes
    - `initial_state`: All machines on connect
    
  version: 2.0.0
  contact:
    name: AgriTrack Team
    url: https://github.com/KD-3030/AgriTrack
servers:
  - url: http://localhost:3001
    description: Local development server
...
paths:
...
  /api/v1/bookings/farmer/{id}:
    get:
      tags:
        - Bookings
      summary: Get farmer's bookings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
...
  /api/v1/farmers:
    get:
      tags:
        - Farmers
      summary: List all farmers
      description: Get a list of all registered farmers (admin only)
      parameters:
        - name: district
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: verified
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of farmers
          content:
            application/json:
              schema:
                type: object
                properties:
                  farmers:
                    type: array
                    items:
                      $ref: '#/components/schemas/FarmerProfile'
                  total:
                    type: integer
    post:
      tags:
        - Farmers
      summary: Register a new farmer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FarmerRegistration'
      responses:
        '201':
          description: Farmer registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  farmer:
                    $ref: '#/components/schemas/FarmerProfile'
        '409':
          description: Farmer already exists
  /api/v1/farmers/{id}:
    get:
      tags:
        - Farmers
      summary: Get farmer profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Farmer profile with fields and service history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FarmerProfile'
    put:
      tags:
        - Farmers
      summary: Update farmer profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FarmerProfileUpdate'
      responses:
        '200':
          description: Profile updated
  /api/v1/farmers/phone/{phone}:
    get:
      tags:
        - Farmers
      summary: Get farmer by phone number
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Farmer profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FarmerProfile'
        '404':
          description: Farmer not found
  /api/v1/farmers/{id}/verify:
    post:
      tags:
        - Farmers
      summary: Verify farmer (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Farmer verified
  /api/v1/farmers/{id}/fields:
    get:
      tags:
        - Farmers
      summary: Get farmer's fields
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  fields:
                    type: array
                    items:
                      $ref: '#/components/schemas/FarmerField'
    post:
      tags:
        - Farmers
      summary: Add a new field
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FarmerFieldRequest'
      responses:
        '201':
          description: Field added
  /api/v1/farmers/{farmerId}/fields/{fieldId}:
    put:
      tags:
        - Farmers
      summary: Update a field
      parameters:
        - name: farmerId
          in: path
          required: true
          schema:
            type: string
        - name: fieldId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FarmerFieldRequest'
      responses:
        '200':
          description: Field updated
    delete:
      tags:
        - Farmers
      summary: Delete a field
      parameters:
        - name: farmerId
          in: path
          required: true
          schema:
            type: string
        - name: fieldId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Field deleted
  /api/v1/farmers/{id}/dashboard:
    get:
      tags:
        - Farmers
      summary: Get farmer dashboard data
      description: Returns profile, stats, recent bookings, and service history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FarmerDashboard'
  /api/v1/farmers/{id}/bookings:
    get:
      tags:
        - Farmers
      summary: Get farmer's bookings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'
  /api/v1/farmers/{id}/transactions:
    get:
      tags:
        - Farmers
      summary: Get farmer's transaction history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/FarmerTransaction'
  /api/v1/farmers/{id}/notifications:
    get:
      tags:
        - Farmers
      summary: Get farmer's notifications
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: unread_only
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/FarmerNotification'
  /api/v1/farmers/{id}/services:
    post:
      tags:
        - Farmers
      summary: Record a service for farmer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRecord'
      responses:
        '201':
          description: Service recorded
  /api/v1/farmers/{id}/services/{serviceId}/feedback:
    post:
      tags:
        - Farmers
      summary: Add feedback for a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: serviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                feedback:
                  type: string
      responses:
        '200':
          description: Feedback added
          description: Dashboard analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Analytics'
...
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: string
        farmer_id:
          type: string
        machine_id:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled]
...
    FarmerProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        full_name:
          type: string
        phone:
          type: string
        alternate_phone:
          type: string
        email:
          type: string
        profile_image_url:
          type: string
        address_line1:
          type: string
        address_line2:
          type: string
        village:
          type: string
        district:
          type: string
        state:
          type: string
        pincode:
          type: string
        farming_experience_years:
          type: integer
        primary_crops:
          type: array
          items:
            type: string
        preferred_language:
          type: string
        is_verified:
          type: boolean
        verification_date:
          type: string
          format: date-time
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    FarmerRegistration:
      type: object
      required:
        - full_name
        - phone
        - district
        - state
      properties:
        full_name:
          type: string
          example: Ranjit Singh
        phone:
          type: string
          example: "+919876543210"
        email:
          type: string
        alternate_phone:
          type: string
        address_line1:
          type: string
        address_line2:
          type: string
        village:
          type: string
          example: Khanna
        district:
          type: string
          example: Ludhiana
        state:
          type: string
          example: Punjab
        pincode:
          type: string
        farming_experience_years:
          type: integer
          example: 10
        primary_crops:
          type: array
          items:
            type: string
          example: ["wheat", "rice", "cotton"]
        preferred_language:
          type: string
          example: hindi
    FarmerProfileUpdate:
      type: object
      properties:
        full_name:
          type: string
        alternate_phone:
          type: string
        email:
          type: string
        address_line1:
          type: string
        address_line2:
          type: string
        village:
          type: string
        pincode:
          type: string
        farming_experience_years:
          type: integer
        primary_crops:
          type: array
          items:
            type: string
        preferred_language:
          type: string
    FarmerField:
      type: object
      properties:
        id:
          type: string
          format: uuid
        farmer_id:
          type: string
          format: uuid
        name:
          type: string
          example: Main Field
        area_acres:
          type: number
          example: 12.5
        village:
          type: string
        district:
          type: string
        state:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        boundary_coords:
          type: array
          items:
            type: object
            properties:
              lat:
                type: number
              lng:
                type: number
        soil_type:
          type: string
          enum: [alluvial, black, red, laterite, sandy, loamy]
        irrigation_type:
          type: string
          enum: [canal, tubewell, rainfed, drip, sprinkler]
        current_crop:
          type: string
        last_harvest_date:
          type: string
          format: date
        next_sowing_date:
          type: string
          format: date
        is_active:
          type: boolean
    FarmerFieldRequest:
      type: object
      required:
        - name
        - area_acres
      properties:
        name:
          type: string
        area_acres:
          type: number
        village:
          type: string
        district:
          type: string
        state:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        boundary_coords:
          type: object
        soil_type:
          type: string
        irrigation_type:
          type: string
        current_crop:
          type: string
...
    ServiceRecord:
      type: object
      required:
        - service_type
        - service_date
      properties:
        booking_id:
          type: string
        field_id:
          type: string
        machine_id:
          type: string
        service_type:
          type: string
        service_date:
          type: string
          format: date
        acres_serviced:
          type: number
        hours_worked:
          type: number
        rate_per_acre:
          type: number
        total_amount:
          type: number
        payment_method:
          type: string
...